# src/server/routes/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "version": "1.0.0",
        "title": "UTD Degree Planner and Auditor",
        "description": "API Documentation for UTD Degree Planner and Auditor"
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
...
    "paths": {
...
        "/api/buildDegree/degree": {
            "post": {
                "description": "Create a new degree",
                "summary": "Create a new degree",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name",
                                    "year"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/degree/:name/:year": {
            "get": {
                "description": "Get a degree by name and year",
                "summary": "Get a degree by name and year",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "name",
                        "in": "path"
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "year",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/:degreePlanID": {
            "get": {
                "description": "Get a degree plan by ID",
                "summary": "Get a degree plan by ID",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "degreePlanID",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/": {
            "post": {
                "description": "Create a new degree plan",
                "summary": "Create a new degree plan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userID": {
                                        "type": "string"
                                    },
                                    "degreeID": {
                                        "type": "object",
                                        "properties": {
                                            "degreeName": {
                                                "type": "string"
                                            },
                                            "degreeYear": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "degreeName",
                                            "degreeYear"
                                        ]
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "userID",
                                    "degreeID",
                                    "name"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/addCourse": {
            "post": {
                "description": "Add a course to a degree plan",
                "summary": "Add a course to a degree plan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "degreePlanID": {
                                        "type": "string"
                                    },
                                    "course": {
                                        "type": "object",
                                        "properties": {
                                            "prefix": {
                                                "type": "string"
                                            },
                                            "number": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "prefix",
                                            "number"
                                        ]
                                    }
                                },
                                "required": [
                                    "degreePlanID",
                                    "course"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/removeCourse": {
            "delete": {
                "description": "Remove a course from a degree plan",
                "summary": "Remove a course from a degree plan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "degreePlanCourseID": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "degreePlanCourseID"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/updateCourseSemester": {
            "put": {
                "description": "Update a course's semester in a degree plan",
                "summary": "Update a course's semester in a degree plan",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "degreePlanCourseID": {
                                        "type": "string"
                                    },
                                    "semester": {
                                        "type": [
                                            "object",
                                            "null"
                                        ],
                                        "properties": {
                                            "semesterTerm": {
                                                "type": "string",
                                                "enum": [
                                                    "FALL",
                                                    "SPRING",
                                                    "SUMMER"
                                                ]
                                            },
                                            "semesterYear": {
                                                "type": "string"
                                            }
                                        },
                                        "required": [
                                            "semesterTerm",
                                            "semesterYear"
                                        ]
                                    }
                                },
                                "required": [
                                    "degreePlanCourseID",
                                    "semester"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/:degreePlanID/getAllCourseToRequirementBlockLinks": {
            "get": {
                "description": "Get all course to requirement block links",
                "summary": "Get all course to requirement block links",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "degreePlanID",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/linkCourseToRequirementBlock": {
            "put": {
                "description": "Link a course to a requirement block",
                "summary": "Link a course to a requirement block",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "degreePlanCourseID": {
                                        "type": "string"
                                    },
                                    "blockID": {
                                        "type": "string"
                                    },
                                    "credit": {
                                        "type": "integer",
                                        "minimum": 0
                                    }
                                },
                                "required": [
                                    "degreePlanCourseID",
                                    "blockID",
                                    "credit"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/unlinkCourseFromRequirementBlock": {
            "delete": {
                "description": "Unlink a course from a requirement block",
                "summary": "Unlink a course from a requirement block",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "degreePlanCourseID": {
                                        "type": "string"
                                    },
                                    "blockID": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "degreePlanCourseID",
                                    "blockID"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/addTransferCredit": {
            "post": {
                "description": "Add a transfer credit",
                "summary": "Add a transfer credit",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userID": {
                                        "type": "string"
                                    },
                                    "transferCourseEquivalencyID": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "userID",
                                    "transferCourseEquivalencyID"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/degreePlan/applyTransferCredit": {
            "post": {
                "description": "Apply a transfer credit",
                "summary": "Apply a transfer credit",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userID": {
                                        "type": "string"
                                    },
                                    "degreePlanID": {
                                        "type": "string"
                                    },
                                    "prefix": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "transferCourseEquivalencyID": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "userID",
                                    "degreePlanID",
                                    "prefix",
                                    "number",
                                    "transferCourseEquivalencyID"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}